Combine preconfigured section file creation and visualization layer screenshot generation into a single test file; refactor and remove redundant tests.
