+++ tests/acceptance/test_tracks_add_and_display.py
@@ -3,21 +3,7 @@ from pathlib import Path
 import pytest
 from playwright.sync_api import Page  # type: ignore
 
-from OTAnalytics.application.resources.resource_manager import (
-    ResourceManager,
-    VisualizationLayersKeys,
-    VisualizationOffsetSliderKeys,
-)
-from OTAnalytics.plugin_ui.nicegui_gui.pages.visualization_layers_form.layers_form import (  # noqa
-    MARKER_VISUALIZATION_LAYERS_ALL,
-)
-from OTAnalytics.plugin_ui.visualization.visualization import (
-    ALL,
-    ASSIGNED_TO_FLOWS,
-    INTERSECTING_SECTIONS,
-    NOT_ASSIGNED_TO_FLOWS,
-    NOT_INTERSECTING_SECTIONS,
-)
+from OTAnalytics.application.resources.resource_manager import ResourceManager
 from tests.acceptance.conftest import PLAYWRIGHT_VISIBLE_TIMEOUT_MS, NiceGUITestServer
 from tests.utils.playwright_helpers import (
     capture_and_verify_baseline,
@@ -186,118 +172,6 @@ def test_toggle_intersection_layers(
     # Capture screenshot and verify against reference
     verify_canvas_matches_reference(
         canvas, acceptance_test_data_folder, ALL_TRACKS_FILE_NAME
-    )
-
-
-@pytest.mark.skip(reason="only works in headed right now")
-@pytest.mark.timeout(300)
-@pytest.mark.playwright
-@pytest.mark.usefixtures("external_app")
-def test_generate_canvas_screenshots(
-    external_app: NiceGUITestServer,
-    page: Page,
-    resource_manager: ResourceManager,
-    acceptance_test_data_folder: Path,
-) -> None:
-    """Generate reference screenshots for all visualization layer states."""
-    # Setup: Load tracks with preconfigured file
-    base_url = getattr(external_app, "base_url", "http://127.0.0.1:8080")
-    navigate_to_main_page_with_url(page, base_url)
-    data_dir = Path(__file__).parents[1] / "data"
-    otconfig_path = data_dir / "sections_created_test_file.otconfig"
-    canvas = get_loaded_tracks_canvas_from_otconfig(
-        page, resource_manager, otconfig_path
-    )
-
-    # Helper function to toggle a checkbox and take screenshot
-    def toggle_and_screenshot(
-        checkbox_text: str, filename_base: str, nth: int = 0
-    ) -> None:
-        """Toggle a checkbox on, take screenshot, then toggle off.
-
-        Args:
-            checkbox_text: Text of the checkbox to find
-            filename_base: Base name for screenshot files
-            nth: Which occurrence to use if multiple checkboxes have same text
-        """
-        checkbox = page.get_by_text(checkbox_text, exact=True).nth(nth)
-        checkbox.scroll_into_view_if_needed()
-
-        # Toggle on and screenshot
-        if not checkbox.is_checked():
-            checkbox.click()
-        page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-        canvas.screenshot(path=acceptance_test_data_folder / f"{filename_base}.png")
-
-        # Toggle off (no screenshot)
-        checkbox.click()
-        page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-
-    # 1. Show all tracks (already enabled, just take screenshot)
-    page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-    canvas.screenshot(path=acceptance_test_data_folder / ALL_TRACKS_FILE_NAME)
-
-    # Sections and flows are already loaded from the preconfigured file
-    # Just update flow highlighting
-    page.get_by_text(
-        resource_manager.get(VisualizationLayersKeys.BUTTON_UPDATE_FLOW_HIGHLIGHTING),
-        exact=True,
-    ).click()
-    page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-
-    # Turn off "Show all tracks" to prepare for next screenshots
-    all_tracks_checkbox = page.get_by_test_id(MARKER_VISUALIZATION_LAYERS_ALL)
-    all_tracks_checkbox.click()
-    page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-
-    # 2. Show offset ändern und zurücksetzen (Update with section offset button)
-    # This button may be disabled if preconditions aren't met, so skip if not enabled
-    offset_button = page.get_by_text(
-        resource_manager.get(VisualizationOffsetSliderKeys.BUTTON_UPDATE_OFFSET),
-        exact=True,
-    )
-    if offset_button.count() > 0:
-        offset_button.scroll_into_view_if_needed()
-        canvas.screenshot(path=acceptance_test_data_folder / "offset_before.png")
-        # Only click if enabled
-        if not offset_button.is_disabled():
-            offset_button.click()
-            page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-            canvas.screenshot(path=acceptance_test_data_folder / "offset_after.png")
-
-    # 3-11. Toggle all visualization layers and take screenshots
-    # Deselect "Show all tracks" before testing individual highlight options
-    all_tracks_checkbox = page.get_by_test_id(MARKER_VISUALIZATION_LAYERS_ALL)
-    if all_tracks_checkbox.is_checked():
-        all_tracks_checkbox.click()
-        page.wait_for_timeout(PLAYWRIGHT_VISIBLE_TIMEOUT_MS)
-
-    # Show tracks group (nth=0 means first occurrence in "Show tracks" section)
-    toggle_and_screenshot(
-        INTERSECTING_SECTIONS, "highlight_tracks_intersecting_sections", nth=0
-    )
-    toggle_and_screenshot(
-        NOT_INTERSECTING_SECTIONS, "highlight_tracks_not_intersecting_sections", nth=0
-    )
-    toggle_and_screenshot(
-        ASSIGNED_TO_FLOWS, "highlight_tracks_assigned_to_flows", nth=0
-    )
-    toggle_and_screenshot(
-        NOT_ASSIGNED_TO_FLOWS, "highlight_tracks_not_assigned_to_flows", nth=0
-    )
-
-    # Show start and end points group
-    # (nth=1 means second occurrence in "Show start and end points" section)
-    toggle_and_screenshot(
-        INTERSECTING_SECTIONS, "start_end_intersecting_sections", nth=1
-    )
-    toggle_and_screenshot(
-        NOT_INTERSECTING_SECTIONS, "start_end_not_intersecting_sections", nth=1
-    )
-    toggle_and_screenshot(ALL, "start_end_all", nth=1)
-    toggle_and_screenshot(ASSIGNED_TO_FLOWS, "start_end_assigned_to_flows", nth=1)
-    toggle_and_screenshot(
-        NOT_ASSIGNED_TO_FLOWS, "start_end_not_assigned_to_flows", nth=1
     )
